consolidate/remove unutilized code
