+++ src/CMakeLists.txt
@@ -116,8 +116,7 @@ add_subdirectory(checkpoints)
 add_subdirectory(cryptonote_basic)
 add_subdirectory(cryptonote_core)
 add_subdirectory(cryptonote_protocol)
-add_subdirectory(komodo)
-add_subdirectory(komodo_notary_server)
+add_subdirectory(notary_server)
 add_subdirectory(multisig)
 if(NOT IOS)
   add_subdirectory(blockchain_db)
