+++ src/blockchain_utilities/blockchain_blackball.cpp
@@ -37,7 +37,7 @@
 #include "cryptonote_core/blockchain.h"
 #include "blockchain_db/blockchain_db.h"
 #include "blockchain_db/db_types.h"
-#include "komodo/komodo_validation.h"
+#include "cryptonote_core/komodo_validation.h"
 #include "wallet/ringdb.h"
 #include "version.h"
