+++ src/blockchain_utilities/blockchain_usage.cpp
@@ -28,7 +28,6 @@
 
 #include <memory>
 
-#include "komodo/komodo_validation.h"
 #include <boost/range/adaptor/transformed.hpp>
 #include <boost/algorithm/string.hpp>
 #include "common/command_line.h"
@@ -37,6 +36,7 @@
 #include "cryptonote_core/cryptonote_core.h"
 #include "cryptonote_core/blockchain.h"
 #include "cryptonote_core/komodo_notaries.h"
+#include "cryptonote_core/komodo_validation.h"
 #include "blockchain_db/blockchain_db.h"
 #include "blockchain_db/db_types.h"
 #include "version.h"
