+++ src/cryptonote_core/blockchain.cpp
@@ -54,8 +54,8 @@
 #include "warnings.h"
 #include "crypto/hash.h"
 #include "cryptonote_core.h"
-#include "komodo/komodo_validation.h"
 #include "komodo_notaries.h"
+#include "komodo_validation.h"
 #include "ringct/rctSigs.h"
 #include "common/perf_timer.h"
 #include "libhydrogen/hydrogen.h"
