+++ src/cryptonote_protocol/CMakeLists.txt
@@ -48,7 +48,6 @@ monero_add_library(cryptonote_protocol
 target_link_libraries(cryptonote_protocol
         PUBLIC
         p2p
-        komodo
         ${Boost_DATE_TIME_LIBRARY}
         ${Boost_PROGRAM_OPTIONS_LIBRARY}
         ${Boost_SERIALIZATION_LIBRARY}
