+++ src/cryptonote_protocol/cryptonote_protocol_handler.inl
@@ -40,7 +40,7 @@
 #include "cryptonote_basic/cryptonote_format_utils.h"
 #include "profile_tools.h"
 #include "net/network_throttle-detail.hpp"
-#include "komodo_notary_server/notarization_tx_cache.h"
+#include "notary_server/notarization_tx_cache.h"
 
 #undef MONERO_DEFAULT_LOG_CATEGORY
 #define MONERO_DEFAULT_LOG_CATEGORY "net.cn"
