+++ src/daemon/CMakeLists.txt
@@ -95,7 +95,6 @@ target_link_libraries(daemon
     serialization
     daemon_rpc_server
     epee
-    komodo
     ${EVENT_LIBRARY}
     ${SECP256K1_LIBRARY}
     ${BTC_LIBRARY}
