+++ src/gen_multisig/CMakeLists.txt
@@ -39,7 +39,6 @@ target_link_libraries(gen_multisig
     cncrypto
     common
     epee
-    komodo
     ${EPEE_READLINE}
     ${Boost_CHRONO_LIBRARY}
     ${Boost_PROGRAM_OPTIONS_LIBRARY}
