+++ src/notary_server/CMakeLists.txt
@@ -1,26 +1,26 @@
-set(komodo_notary_server_sources
+set(notary_server_sources
   notary_server.cpp
   notarization_tx_cache.cpp)
 
-set(komodo_notary_server_headers)
+set(notary_server_headers)
 
-set(komodo_notary_server_private_headers
+set(notary_server_private_headers
   notary_server.h
   notary_server_error_codes.h
   notary_server_commands_defs.h
   notarization_tx_cache.h
   notarization_tx_container.h)
 
-monero_private_headers(komodo_notary_server
-  ${komodo_notary_server_private_headers})
-monero_add_executable(komodo_notary_server
-  ${komodo_notary_server_sources}
-  ${komodo_notary_server_headers}
-  ${komodo_notary_server_private_headers})
+monero_private_headers(notary_server
+  ${notary_server_private_headers})
+monero_add_executable(notary_server
+  ${notary_server_sources}
+  ${notary_server_headers}
+  ${notary_server_private_headers})
 
-add_backward(komodo_notary_server)
+add_backward(notary_server)
 
-target_link_libraries(komodo_notary_server
+target_link_libraries(notary_server
   PRIVATE
     wallet
     epee
@@ -30,7 +30,6 @@ target_link_libraries(komodo_notary_server
     cncrypto
     p2p
     rpc
-    komodo
     common
     version
     ${BTC_LIBRARY}
@@ -43,7 +42,7 @@ target_link_libraries(komodo_notary_server
     ${CMAKE_THREAD_LIBS_INIT}
     ${EXTRA_LIBRARIES})
 
-set_property(TARGET komodo_notary_server
+set_property(TARGET notary_server
   PROPERTY
     OUTPUT_NAME "blur-notary-server-rpc")
-install(TARGETS komodo_notary_server DESTINATION bin)
+install(TARGETS notary_server DESTINATION bin)
