+++ src/rpc/CMakeLists.txt
@@ -108,7 +108,6 @@ target_link_libraries(rpc
     common
     cryptonote_core
     cryptonote_protocol
-    komodo
     epee
     ${Boost_REGEX_LIBRARY}
     ${Boost_THREAD_LIBRARY}
